docs: erros swagger
